+++ src/App.vue
@@ -1,5 +1,16 @@
-<script setup></script>
+<script lang="ts" setup>
+import { useMarkmap } from '@/stores/markmap'
+const markmapStore = useMarkmap()
+</script>
 
-<template></template>
+<template>
+  <div
+    class="relative"
+    :class="!markmapStore.manual && markmapStore.theme === 'dark' ? 'dark' : ''"
+  >
+    <Markmap />
+    <Help />
+  </div>
+</template>
 
 <style></style>