+++ src/funcs.ts
@@ -6,6 +6,8 @@ import replaceAsync from 'string-replace-async'
 import ellipsis from 'text-ellipsis'
 import TurndownService from 'turndown'
 
+import { useHelp } from '@/stores/help'
+
 const settingsVersion = 'v2'
 export const defaultSettings = {
   keyBindings: {
@@ -168,8 +170,8 @@ export function addToolbar(mm) {
       'M18 10a8 8 0 11-16 0 8 8 0 0116 0zm-7-4a1 1 0 11-2 0 1 1 0 012 0zM9 9a1 1 0 000 2v3a1 1 0 001 1h1a1 1 0 100-2v-3a1 1 0 00-1-1H9z'
     ),
     onClick: async () => {
-      // @ts-ignore
-      Alpine.store('markmap').toggleHelp()
+      const helpStore = useHelp()
+      helpStore.toggleHelp()
     },
   })
   toolbar.attach(mm)